requirements.txt
fastapi
uvicorn
langchain
openai
google-cloud-translate
gtts
python-multipart 